% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Download a file or directory tracked by DVC or by Git into the current working directory.}
\usage{
get(url = "", path = "", out = "", rev = "")
}
\arguments{
\item{url}{The url of the repo}

\item{path}{The path or filename to download}

\item{out}{Specify a path to the desired location in the workspace to place
the downloaded file or directory}

\item{rev}{commit hash, branch or tag name, etc. (any Git revision) of the repository
to download the file or directory from.}
}
\description{
Documentation: <https://dvc.org/doc/command-reference/get>
}
\examples{
  dvcr::get("https://github.com/iterative/example-get-started","model.pkl")
  dvcr::get("https://github.com/iterative/example-get-started","model.pkl",out="data")
}
