% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Download a file or directory tracked by another DVC or Git repository into the workspace,
and track it (an import .dvc file is created).}
\usage{
import(url = "", path = "", out = "", rev = "")
}
\arguments{
\item{url}{The url of the repo}

\item{path}{The path or filename to download}

\item{out}{Specify a path to the desired location in the workspace to place the downloaded file or directory}

\item{rev}{commit hash, branch or tag name, etc. (any Git revision) of the repository
to download the file or directory from.}
}
\description{
Documentation: <https://dvc.org/doc/command-reference/import>
You need to have your project as a dvc repo using dvcr::use_dvc prior
to using this command
}
\examples{
  dvcr::import("git@github.com:iterative/example-get-started","data/data.xml")
  dvcr::import("git@github.com:iterative/example-get-started","data/data.xml",out="data")
}
