% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Download tracked files or directories from remote storage.}
\usage{
pull(targets = c(), remote = FALSE)
}
\arguments{
\item{targets}{Limit command scope to these tracked files/directories, .dvc files, or stage names.}

\item{remote}{Remote storage to pull from.}
}
\description{
Documentation: <https://man.dvc.org/push>
}
