% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Upload tracked files or directories to remote storage.}
\usage{
push(targets = c(), remote = "")
}
\arguments{
\item{targets}{Limit command scope to these tracked files/directories, .dvc files, or stage names.}

\item{remote}{Remote storage to push to.}
}
\description{
Documentation: <https://man.dvc.org/push>
}
