% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_add.R
\name{remote_add}
\alias{remote_add}
\title{Add a new data remote.}
\usage{
remote_add(name, url, default = TRUE, repo = here::here())
}
\arguments{
\item{name}{Name of the remote}

\item{url}{URL. See full list of supported urls at <https://man.dvc.org/remote>}

\item{default}{Set as default remote.}

\item{repo}{Path to root of repository (project root by default)}
}
\description{
Add a new data remote.
}
