% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dvc.R
\name{use_dvc}
\alias{use_dvc}
\title{Initialize a DVC project.}
\usage{
use_dvc(repo = here::here())
}
\arguments{
\item{repo}{Path to root of repository (project root by default)}
}
\description{
Initialize DVC in the current directory. Expects directory
to be a Git repository unless --no-scm option is specified.
Documentation: <https://man.dvc.org/init>
}
